feat: a wrapper component for a form with navigation
